fix: correct logic errors found during codebase audit
- Fix bitwise AND that should be OR in PER OID encoding (per.go:166),
  which corrupted all Object Identifier bytes in MCS/GCC negotiation

- Fix nil pointer dereference in BitmapData.Deserialize when server
  sends compressed bitmap with header (update_events.go:178); the
  CompressedDataHeader pointer was never initialized before use

- Add bounds check in handleWave before slicing body[12:] to prevent
  panic on malformed SNDC_WAVE PDUs (audio.go:245)

- Add bounds check in NLA ASN.1 parsing before accessing spki[offset+1]
  to prevent panic on truncated SubjectPublicKeyInfo (nla.go:336-340)

- Fix uint32 overflow in display_control.go resolution validation;
  width*height could silently wrap, bypassing the max area check.
  Now uses uint64 arithmetic (display_control.go:228)

- Update version test to match bumped version (1.0.1)

+++ internal/rdp/display_control.go
@@ -225,7 +225,7 @@ func (h *DisplayControlHandler) RequestResize(width, height uint32) error {
 	// Validate against server capabilities
 	if h.caps != nil {
 		maxArea := h.caps.MaxMonitorAreaSize
-		if maxArea > 0 && width*height > maxArea {
+		if maxArea > 0 && uint64(width)*uint64(height) > uint64(maxArea) {
 			h.mu.Unlock()
 			return fmt.Errorf("requested resolution %dx%d exceeds server max area %d", width, height, maxArea)
 		}
